+++ site/src/pages/WorkspaceSettingsPage/WorkspaceSharingPage/UserOrGroupAutocomplete.tsx
@@ -4,10 +4,10 @@ import type { Group, User } from "api/typesGenerated";
 import { Autocomplete } from "components/Autocomplete/Autocomplete";
 import { AvatarData } from "components/Avatar/AvatarData";
 import { Check } from "lucide-react";
+import { getGroupSubtitle, isGroup } from "modules/groups";
 import { type FC, useState } from "react";
 import { keepPreviousData, useQuery } from "react-query";
 import { prepareQuery } from "utils/filters";
-import { getGroupSubtitle, isGroup } from "utils/groups";
 
 type AutocompleteOption = User | Group;
 export type UserOrGroupAutocompleteValue = AutocompleteOption | null;
